Move UI components into folder
