+++ src/components/selection/candidateSelect.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Button from '../button';
+import Button from '../ui/button';
 
 const CandidateSelect = ({ candidate }) => (
   <Button id={candidate.id}>{candidate.name}</Button>
