+++ src/components/ui/button.js
@@ -2,7 +2,7 @@ import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 
-import { Context } from '../state/reducer';
+import { Context } from '../../state/reducer';
 
 const StyledButton = styled.button`
   border: ${props =>
